requirements.txt
streamlit==1.29.0
pillow==10.4.0
exifread==3.0.0
pandas==2.1.3
python-dotenv==1.0.0
pymongo==4.6.0
dnspython==2.4.2
web3==7.13.0
eth-account==0.13.7
py-solc-x==1.12.2
stegano==0.11.2
pyzbar==0.1.9
opencv-python-headless==4.8.0.76
numpy==1.24.3
scikit-image==0.21.0